--- configs/emmet.lua
@@ -1,27 +1,0 @@
-local lspconfig = require "lspconfig"
-local capabilities = vim.lsp.protocol.make_client_capabilities()
-capabilities.textDocument.completion.completionItem.snippetSupport = true
-
-return function()
-  lspconfig.emmet_ls.setup {
-    -- on_attach = on_attach,
-    capabilities = capabilities,
-    filetypes = {
-      "html",
-      "typescriptreact",
-      "javascriptreact",
-      "css",
-      "sass",
-      "scss",
-      "less",
-    },
-    init_options = {
-      html = {
-        options = {
-          -- For possible options, see: https://github.com/emmetio/emmet/blob/master/src/config.ts#L79-L267
-          ["bem.enabled"] = true,
-        },
-      },
-    },
-  }
-end
