--- plugins.lua
@@ -1,225 +1,0 @@
-local overrides = require "custom.configs.overrides"
-
----@type NvPluginSpec[]
-local plugins = {
-  -- Override plugin definition options
-  {
-    "neovim/nvim-lspconfig",
-    dependencies = {
-      -- format & linting
-      {
-        "nvimtools/none-ls.nvim",
-        config = function()
-          require "custom.configs.null-ls"
-        end,
-      },
-    },
-    config = function()
-      require "plugins.configs.lspconfig"
-      require "custom.configs.lspconfig"
-    end, -- Override to setup mason-lspconfig
-  },
-
-  -- override plugin configs
-  {
-    "williamboman/mason.nvim",
-    opts = overrides.mason,
-  },
-
-  {
-    "nvim-treesitter/nvim-treesitter",
-    opts = overrides.treesitter,
-  },
-
-  {
-    "nvim-tree/nvim-tree.lua",
-    opts = overrides.nvimtree,
-  },
-
-  -- Install a plugin
-  {
-    "max397574/better-escape.nvim",
-    event = "InsertEnter",
-    config = function()
-      require("better_escape").setup()
-    end,
-  },
-  {
-    "windwp/nvim-ts-autotag",
-    event = "InsertEnter",
-    config = function()
-      require("nvim-ts-autotag").setup()
-    end,
-  },
-  {
-    "dkarter/bullets.vim",
-    event = "InsertEnter",
-    config = function()
-      vim.g.bullets_enabled_file_types = {
-        "markdown",
-        "text",
-        "gitcommit",
-        "scratch",
-      }
-    end,
-  },
-  {
-    "NeogitOrg/neogit",
-    config = true,
-    dependencies = {
-      "nvim-lua/plenary.nvim",  -- required
-      "sindrets/diffview.nvim", -- optional - Diff integration
-
-      -- Only one of these is needed, not both.
-      "nvim-telescope/telescope.nvim", -- optional
-      "ibhagwan/fzf-lua",              -- optional
-    },
-    keys = {
-      { "<leader>gg", ":Neogit<CR>", desc = "Neogit toggle" },
-    },
-  },
-  {
-    "kylechui/nvim-surround",
-    event = "InsertEnter",
-    version = "*", -- Use for stability; omit to use `main` branch for the latest features
-    config = function()
-      require("nvim-surround").setup {
-        -- Configuration here, or leave empty to use defaults
-      }
-    end,
-  },
-  {
-    "ThePrimeagen/harpoon",
-    event = "InsertEnter",
-    keys = {
-      {
-        "<leader>ml",
-        function()
-          require("harpoon.ui").toggle_quick_menu()
-        end,
-        desc = "list marks",
-      },
-      {
-        "<leader>mt",
-        function()
-          require("harpoon.term").gotoTerminal()
-        end,
-        desc = "term",
-      },
-      {
-        "<leader>mm",
-        function()
-          require("harpoon.mark").add_file()
-        end,
-        desc = "Show marks",
-      },
-      {
-        "<leader>mn",
-        function()
-          require("harpoon.ui").nav_next()
-        end,
-        desc = "Next mark",
-      },
-      {
-        "<leader>mb",
-        function()
-          require("harpoon.ui").nav_prev()
-        end,
-        desc = "Prev mark",
-      },
-    },
-  },
-  {
-    "wakatime/vim-wakatime",
-    event = "InsertEnter",
-  },
-  { "nvim-ts-autotag", opts = {} },
-  {
-    "folke/todo-comments.nvim",
-    event = "LspAttach",
-    dependencies = { "nvim-lua/plenary.nvim" },
-    opts = require "custom.configs.comments",
-  },
-  {
-    "epwalsh/obsidian.nvim",
-    version = "*", -- recommended, use latest release instead of latest commit
-    lazy = true,
-    ft = "markdown",
-    dependencies = {
-      -- Required.
-      "nvim-lua/plenary.nvim",
-    },
-    opts = overrides.obsidian,
-  },
-
-  -- Personal plugins
-  {
-    "marco-souza/ncm.nvim",
-    cmd = "Config",
-    dir = "~/workspace/marco-souza/ncm.nvim/",
-    dependencies = { "nvim-lua/plenary.nvim", "nvim-telescope/telescope-ui-select.nvim" },
-    config = true,
-  },
-  {
-    "marco-souza/ollero.nvim",
-    event = "VeryLazy",
-    dependencies = {
-      -- {
-      --   "marco-souza/term.nvim",
-      --   dir = "~/workspace/marco-souza/term.nvim/",
-      -- },
-      "marco-souza/term.nvim",
-      "nvim-telescope/telescope-ui-select.nvim",
-    },
-    dir = "~/workspace/marco-souza/ollero.nvim/",
-    opts = {},
-  },
-  {
-    "nvim-neorg/neorg",
-    ft = "norg",
-    build = ":Neorg sync-parsers",
-    dependencies = { "nvim-lua/plenary.nvim" },
-    opts = {
-      load = {
-        ["core.defaults"] = {},  -- Loads default behaviour
-        ["core.concealer"] = {}, -- Adds pretty icons to your documents
-        ["core.dirman"] = {      -- Manages Neorg workspaces
-          config = {
-            workspaces = {
-              notes = "~/workspace/marco-souza/notes",
-            },
-          },
-        },
-      },
-    },
-  },
-
-  -- TODO: waiting for https://github.com/huggingface/llm.nvim/pull/68
-  -- {
-  --   "noahbald/llm.nvim",
-  --   event = "LspAttach",
-  --   version = "main",
-  --   opts = {
-  --     -- model = "http://localhost:11434/api/generate",
-  --     model = "http://localhost:11434/api/generate",
-  --     adaptor = "ollama",
-  --     request_body = { model = "codellama:7b-code" },
-  --   },
-  -- },
-
-  -- To make a plugin not be loaded
-  -- {
-  --   "NvChad/nvim-colorizer.lua",
-  --   enabled = false
-  -- },
-
-  -- All NvChad plugins are lazy-loaded by default
-  -- For a plugin to be loaded, you will need to set either `ft`, `cmd`, `keys`, `event`, or set `lazy = false`
-  -- If you want a plugin to load on startup, add `lazy = false` to a plugin spec, for example
-  -- {
-  --   "mg979/vim-visual-multi",
-  --   lazy = false,
-  -- }
-}
-
-return plugins
