+++ plugins/emmet.lua
@@ -1,0 +1,23 @@
+return {
+  {
+    "aca/emmet-ls",
+    event = "InsertEnter",
+    config = function()
+      local lspconfig = require "lspconfig"
+      local capabilities = vim.lsp.protocol.make_client_capabilities()
+      capabilities.textDocument.completion.completionItem.snippetSupport = true
+
+      lspconfig.emmet_ls.setup {
+        capabilities = capabilities,
+        init_options = {
+          html = {
+            options = {
+              -- For possible options, see: https://github.com/emmetio/emmet/blob/master/src/config.ts#L79-L267
+              ["bem.enabled"] = true,
+            },
+          },
+        },
+      }
+    end,
+  },
+}
