+++ plugins/obsidian.lua
@@ -1,0 +1,15 @@
+return {
+  "epwalsh/obsidian.nvim",
+  version = "*", -- recommended, use latest release instead of latest commit
+  lazy = true,
+  ft = "markdown",
+  dependencies = {
+    -- Required.
+    "nvim-lua/plenary.nvim",
+  },
+  opts = {
+    workspaces = {
+      { name = "notes", path = "~/w/marco-souza/notes" },
+    },
+  },
+}
